pyproject.toml
[project]
name = "ChronoTagger"
version = "0.1.0"
description = "A tool to add dates to images based on filename patterns."
authors = [
    {name = "Carlos Larriu", email = "larriucarlos@gmail.com"},
]
dependencies = [
    "Pillow>=10.4.0",
    "piexif>=1.1.3",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
